Also extract a HttpBase step class
To be used as base class for headless browser steps.
